Port Request to Typescript
